fixes relatorios rastreios atencao farmaceutica
