+++ src/pages/Reports/Reports.vue
@@ -103,7 +103,7 @@ import Mmia from 'components/Reports/ClinicManagement/Mmia.vue';
 import LinhasTerapeuticasUsadas from 'components/Reports/ClinicManagement/LinhasUsadas.vue';
 import SegundasLinhas from 'components/Reports/ClinicManagement/SegundasLinhas.vue';
 import PatientsAbandonment from 'components/Reports/ClinicManagement/PatientsAbandonment.vue';
-import PatientsAbandonmentReturned from 'components/Reports/ClinicManagement/PatientsAbandonmentReturned.vue';
+// import PatientsAbandonmentReturned from 'components/Reports/ClinicManagement/PatientsAbandonmentReturned.vue';
 import ReferredPatients from 'components/Reports/ReferralManagement/ReferredPatients.vue';
 import ReferredBackPatients from 'components/Reports/ReferralManagement/ReferredBackPatients.vue';
 import ReferredPatientDispenseHistory from 'components/Reports/ReferralManagement/ReferredPatientDispenseHistory.vue';
@@ -128,7 +128,7 @@ import { useLoading } from 'src/composables/shared/loading/loading';
 import TBScreening from 'components/Reports/Patient/TBScreening.vue';
 import RAMScreening from 'components/Reports/Patient/RAMScreening.vue';
 import PacksByDrugBottles from 'src/components/Reports/monitoring/PacksByDrugBottles.vue';
-import PatientWithPregnancyScreening from 'src/components/Reports/monitoring/PatientsWithPregnancyScreening.vue';
+import PatientWithPregnancyScreening from 'src/components/Reports/Patient/PatientsWithPregnancyScreening.vue';
 import NotSynchronizedPackToServer from 'components/Reports/monitoring/NotSyncronizedPacksToServer.vue';
 import PatientsMonitoredForAdherence from 'components/Reports/Patient/PatientsMonitoredForAdherence.vue';
 // NOVOS REPORTS COM REUTILIZACAO DE CONTROLLER
@@ -155,7 +155,6 @@ const componentsList = {
   LinhasTerapeuticasUsadas,
   SegundasLinhas,
   PatientsAbandonment,
-  PatientsAbandonmentReturned,
   ReferredBackPatients,
   ReferredPatients,
   ReferredPatientDispenseHistory,
