+++ src/services/api/report/mobile/PatientsWithScreeningMobileService.ts
@@ -1,7 +1,7 @@
 import ReportDatesParams from 'src/services/reports/ReportDatesParams';
 import patientVisitService from '../../patientVisit/patientVisitService';
 import moment from 'moment';
-import PatientsWithPregnancyScreening from 'src/stores/models/report/monitoring/PatientsWithPregnancyScreening';
+import PatientsWithScreeningReport from 'src/stores/models/report/patient/PatientsWithScreeningReport';
 import db from 'src/stores/dexie';
 import { v4 as uuidv4 } from 'uuid';
 import drugService from '../../drugService/drugService';
@@ -10,11 +10,11 @@ import patientServiceIdentifierService from '../../patientServiceIdentifier/pati
 import { useDateUtils } from 'src/composables/shared/dateUtils/dateUtils';
 import clinicService from '../../clinicService/clinicService';
 
-const patientsWithPregnancyScreening = PatientsWithPregnancyScreening.entity;
+const patientsWithScreeningReport = PatientsWithScreeningReport.entity;
 const { idadeCalculator } = useDateUtils();
 
 export default {
-  async getDataLocalDb(params) {
+  async getDataLocalDbForPregnancyScreening(params: any) {
     const reportParams = ReportDatesParams.determineStartEndDate(params);
 
     console.log(reportParams);
@@ -27,8 +27,7 @@ export default {
       );
     console.log(patientVisitList);
     for (const patientVisit of patientVisitList) {
-      const patientsWithPregnancyScreening =
-        new PatientsWithPregnancyScreening();
+      const patientsWithPregnancyScreening = new PatientsWithScreeningReport();
       const endDate = moment(params.endDate).format('YYYY-MM-DD');
       const startDate = moment(params.startDate).format('YYYY-MM-DD');
       patientsWithPregnancyScreening.reportId = reportParams.id;
@@ -66,7 +65,7 @@ export default {
         if (patientVisit.pregnancyScreenings[0].pregnant === true) {
           patientsWithPregnancyScreening.isPregnant = 'Sim';
         } else {
-          patientsWithPregnancyScreening.isPregnant = 'Nao';
+          patientsWithPregnancyScreening.isPregnant = 'Não';
         }
 
         this.localDbAddOrUpdate(patientsWithPregnancyScreening);
@@ -75,7 +74,7 @@ export default {
     }
   },
 
-  async getDataLocalDbMonitoredForAdherence(params) {
+  async getDataLocalDbMonitoredForAdherence(params: any) {
     const reportParams = ReportDatesParams.determineStartEndDate(params);
 
     console.log(reportParams);
@@ -88,8 +87,7 @@ export default {
       );
     console.log(patientVisitList);
     for (const patientVisit of patientVisitList) {
-      const patientsWithPregnancyScreening =
-        new PatientsWithPregnancyScreening();
+      const patientsWithPregnancyScreening = new PatientsWithScreeningReport();
       const endDate = moment(params.endDate).format('YYYY-MM-DD');
       const startDate = moment(params.startDate).format('YYYY-MM-DD');
       patientsWithPregnancyScreening.reportId = reportParams.id;
@@ -130,14 +128,130 @@ export default {
     }
   },
 
+  async getDataLocalDbForTBScreening(params: any) {
+    const reportParams = ReportDatesParams.determineStartEndDate(params);
+    console.log(reportParams);
+    const patientVisitList =
+      await patientVisitService.getLocalDbPatientVisitsBetweenDatesWithTBScreening(
+        params.startDate,
+        params.endDate
+      );
+
+    for (const patientVisit of patientVisitList) {
+      const tbScreening = patientVisit.tbScreenings[0];
+      const tbScreeningReport = new PatientsWithScreeningReport();
+
+      const patient = await patientService.getPatientByIdMobile(
+        patientVisit.patient.id
+      );
+
+      let identifier = patient.identifiers[0];
+
+      if (!identifier) {
+        const idents =
+          await patientServiceIdentifierService.getAllMobileByPatientId(
+            patientVisit.patient.id
+          );
+        identifier = idents[0];
+      }
+      tbScreeningReport.id = uuidv4();
+      tbScreeningReport.nid = identifier.value;
+      tbScreeningReport.name =
+        patient.firstNames +
+        ' ' +
+        patient.middleNames +
+        ' ' +
+        patient.lastNames;
+      tbScreeningReport.age = idadeCalculator(patient.dateOfBirth);
+      tbScreeningReport.gender = patient.gender;
+      tbScreeningReport.dateRegister = patientVisit.visitDate;
+      tbScreeningReport.clinic = clinicService.getById(
+        patientVisit.clinic.id
+      ).clinicName;
+      tbScreeningReport.reportId = reportParams.id;
+      tbScreeningReport.year = reportParams.year;
+      tbScreeningReport.endDate = reportParams.endDate;
+      tbScreeningReport.startDate = reportParams.startDate;
+
+      if (
+        tbScreening.cough === true ||
+        tbScreening.fever === true ||
+        tbScreening.losingWeight === true ||
+        tbScreening.sweating === true ||
+        tbScreening.fatigueOrTirednessLastTwoWeeks === true
+      ) {
+        tbScreeningReport.wasTBScreened = 'Sim';
+      } else {
+        tbScreeningReport.wasTBScreened = 'Não';
+      }
+
+      this.localDbAddOrUpdate(tbScreeningReport);
+    }
+  },
+
+  async getDataLocalDbRAM(params: any) {
+    const reportParams = ReportDatesParams.determineStartEndDate(params);
+    console.log(reportParams);
+    const patientVisitList =
+      await patientVisitService.getLocalDbPatientVisitsBetweenDatesWithRAMScreening(
+        params.startDate,
+        params.endDate
+      );
+
+    for (const patientVisit of patientVisitList) {
+      const ramScreening = patientVisit.tbScreenings[0];
+      const ramScreeningReport = new PatientsWithScreeningReport();
+
+      const patient = await patientService.getPatientByIdMobile(
+        patientVisit.patient.id
+      );
+
+      let identifier = patient.identifiers[0];
+
+      if (!identifier) {
+        const idents =
+          await patientServiceIdentifierService.getAllMobileByPatientId(
+            patientVisit.patient.id
+          );
+        identifier = idents[0];
+      }
+      ramScreeningReport.id = uuidv4();
+      ramScreeningReport.nid = identifier.value;
+      ramScreeningReport.name =
+        patient.firstNames +
+        ' ' +
+        patient.middleNames +
+        ' ' +
+        patient.lastNames;
+      ramScreeningReport.age = idadeCalculator(patient.dateOfBirth);
+      ramScreeningReport.gender = patient.gender;
+      ramScreeningReport.dateRegister = patientVisit.visitDate;
+      ramScreeningReport.clinic = clinicService.getById(
+        patientVisit.clinic.id
+      ).clinicName;
+      ramScreeningReport.reportId = reportParams.id;
+      ramScreeningReport.year = reportParams.year;
+      ramScreeningReport.endDate = reportParams.endDate;
+      ramScreeningReport.startDate = reportParams.startDate;
+
+      if (ramScreening.adverseReactionMedicine === true) {
+        ramScreeningReport.wasRAMScreened = 'Sim';
+      } else {
+        ramScreeningReport.wasRAMScreened = 'Não';
+      }
+
+      this.localDbAddOrUpdate(ramScreeningReport);
+    }
+  },
+
   localDbAddOrUpdate(data: any) {
-    return db[patientsWithPregnancyScreening].add(data).catch((error: any) => {
+    return db[patientsWithScreeningReport].add(data).catch((error: any) => {
       console.log(error);
     });
   },
 
   async localDbGetAllByReportId(reportId: any) {
-    return db[patientsWithPregnancyScreening]
+    return db[patientsWithScreeningReport]
       .where('reportId')
       .equalsIgnoreCase(reportId)
       .toArray()
