+++ src/services/reports/Patients/RAMScreening.ts
@@ -6,7 +6,8 @@ import { MOHIMAGELOG } from 'src/assets/imageBytes.ts';
 import * as ExcelJS from 'exceljs';
 import { useSystemUtils } from 'src/composables/shared/systemUtils/systemUtils';
 import clinicService from 'src/services/api/clinicService/clinicService';
-
+import DownloadFileMobile from 'src/utils/DownloadFileMobile';
+import { fetchFontAsBase64 } from 'src/utils/ReportUtils';
 const { isMobile, isOnline } = useSystemUtils();
 const reportName = 'PacientesComRastreioRAM';
 const logoTitle =
@@ -15,9 +16,11 @@ const title = 'Lista de Pacientes Com rastreio de RAM';
 const fileName = reportName.concat(
   '_' + moment(new Date()).format('DD-MM-YYYY')
 );
+const fontPath = '/src/assets/NotoSans-Regular.ttf';
 
 export default {
   async downloadPDF(province, startDate, endDate, result) {
+    const fontBase64 = await fetchFontAsBase64(fontPath);
     const clinic = clinicService.currClinic();
     const doc = new JsPDF({
       orientation: 'l',
@@ -29,13 +32,16 @@ export default {
     const firstObject = result[0];
     // const totalPagesExp = '{total_pages_count_string}'
 
+    doc.addFileToVFS('NotoSans-Regular.ttf', fontBase64.split(',')[1]);
+    doc.addFont('NotoSans-Regular.ttf', 'NotoSans', 'normal');
+    doc.setFont('NotoSans');
     doc.setProperties({
       title: fileName.concat('.pdf'),
     });
 
     const image = new Image();
-    // image.src = '/src/assets/MoHLogo.png'
-    image.src = 'data:image/png;base64,' + MOHIMAGELOG;
+    image.src = '/src/assets/MoHLogo.png';
+    //image.src = 'data:image/png;base64,' + MOHIMAGELOG;
 
     const headerReport = [
       [
@@ -93,6 +99,7 @@ export default {
 
     autoTable(doc, {
       bodyStyles: {
+        font: 'NotoSans',
         halign: 'left',
         valign: 'middle',
         fontSize: 8,
@@ -108,7 +115,7 @@ export default {
     doc.setFontSize(8);
     doc.text('República de Moçambique ', 16, 28);
     doc.text('Ministério da Saúde ', 20, 32);
-    doc.text('Serviço Nacional de Saúde ', 16, 36);
+    doc.text('Serviço Nacional de Saúde', 16, 36);
     doc.addImage(image, 'png', 28, 15, 10, 10);
 
     const cols = [
@@ -117,8 +124,8 @@ export default {
       'Nome',
       'Sexo',
       'Idade',
-      'Data de criacao',
-      'Unidade Sanitaria',
+      'Data de criação',
+      'Unidade Sanitária',
       'Suspeito',
     ];
 
@@ -145,6 +152,7 @@ export default {
     ord = 0;
     autoTable(doc, {
       bodyStyles: {
+        font: 'NotoSans',
         halign: 'center',
         fontSize: 8,
       },
@@ -184,7 +192,7 @@ export default {
       window.open(doc.output('bloburl'));
     } else {
       const pdfOutput = doc.output();
-      this.downloadFile(fileName, '.pdf', pdfOutput);
+      DownloadFileMobile.downloadFile(fileName, '.pdf', pdfOutput);
     }
     // params.value.loading.loading.hide()
     // return doc.save('HistoricoDeLevantamento.pdf')
@@ -435,70 +443,7 @@ export default {
       saveAs(blob, fileName + fileExtension);
     } else {
       const titleFile = 'RAMScreening.xlsx';
-      saveBlob2File(titleFile, blob);
-      function saveBlob2File(fileName, blob) {
-        const folder = cordova.file.externalRootDirectory + 'Download';
-        //  var folder = 'Download'
-        window.resolveLocalFileSystemURL(
-          folder,
-          function (dirEntry) {
-            createFile(dirEntry, fileName, blob);
-            // $q.loading.hide()
-          },
-          onErrorLoadFs
-        );
-      }
-      function createFile(dirEntry, fileName, blob) {
-        // Creates a new file
-        dirEntry.getFile(
-          fileName,
-          { create: true, exclusive: false },
-          function (fileEntry) {
-            writeFile(fileEntry, blob);
-          },
-          onErrorCreateFile
-        );
-      }
-
-      function writeFile(fileEntry, dataObj) {
-        // Create a FileWriter object for our FileEntry
-        fileEntry.createWriter(function (fileWriter) {
-          fileWriter.onwriteend = function () {
-            console.log('Successful file write...');
-            openFile();
-          };
-
-          fileWriter.onerror = function (error) {
-            console.log('Failed file write: ' + error);
-          };
-          fileWriter.write(dataObj);
-        });
-      }
-      function onErrorLoadFs(error) {
-        console.log(error);
-      }
-
-      function onErrorCreateFile(error) {
-        console.log('errorr: ' + error.toString());
-      }
-      function openFile() {
-        const strTitle = titleFile;
-        console.log('file system 44444: ' + strTitle);
-        const folder =
-          cordova.file.externalRootDirectory + 'Download/' + strTitle;
-        console.log('file system 2222: ' + folder);
-        const documentURL = decodeURIComponent(folder);
-        cordova.plugins.fileOpener2.open(
-          documentURL,
-          'application/vnd.ms-excel',
-          {
-            error: function (e) {
-              console.log('file system open3333366: ' + e + documentURL);
-            },
-            success: function () {},
-          }
-        );
-      }
+      DownloadFileMobile.downloadFile(titleFile, '.xlsx', blob);
     }
   },
   createArrayOfArrayRow(rows) {
@@ -525,70 +470,5 @@ export default {
     rows = [];
 
     return data;
-  },
-  downloadFile(fileName, fileType, blop) {
-    const titleFile = fileName + fileType;
-    console.log('result' + titleFile);
-    saveBlob2File(titleFile, blop);
-    function saveBlob2File(fileName, blob) {
-      const folder = cordova.file.externalRootDirectory + 'Download';
-      //  var folder = 'Download'
-      window.resolveLocalFileSystemURL(
-        folder,
-        function (dirEntry) {
-          createFile(dirEntry, fileName, blob);
-          // $q.loading.hide()
-        },
-        onErrorLoadFs
-      );
-    }
-    function createFile(dirEntry, fileName, blob) {
-      // Creates a new file
-      dirEntry.getFile(
-        fileName,
-        { create: true, exclusive: false },
-        function (fileEntry) {
-          writeFile(fileEntry, blob);
-        },
-        onErrorCreateFile
-      );
-    }
-
-    function writeFile(fileEntry, dataObj) {
-      // Create a FileWriter object for our FileEntry
-      fileEntry.createWriter(function (fileWriter) {
-        fileWriter.onwriteend = function () {
-          console.log('Successful file write...');
-          openFile();
-        };
-
-        fileWriter.onerror = function (error) {
-          console.log('Failed file write: ' + error);
-        };
-        fileWriter.write(dataObj);
-      });
-    }
-    function onErrorLoadFs(error) {
-      console.log(error);
-    }
-
-    function onErrorCreateFile(error) {
-      console.log('errorr: ' + error.toString());
-    }
-    function openFile() {
-      const strTitle = titleFile;
-      console.log('file system 44444: ' + strTitle);
-      const folder =
-        cordova.file.externalRootDirectory + 'Download/' + strTitle;
-      console.log('file system 2222: ' + folder);
-      const documentURL = decodeURIComponent(folder);
-      cordova.plugins.fileOpener2.open(documentURL, 'application/pdf', {
-        error: function (e) {
-          console.log('file system open3333366: ' + e + documentURL);
-        },
-        success: function () {},
-      });
-    }
-    // }
   },
 };
