+++ src/utils/ReportUtils.ts
@@ -1,0 +1,10 @@
+export async function fetchFontAsBase64(url: any) {
+  const response = await fetch(url);
+  const blob = await response.blob();
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onloadend = () => resolve(reader.result);
+    reader.onerror = reject;
+    reader.readAsDataURL(blob);
+  });
+}
